+++ internal/service/service_test.go
@@ -19,10 +19,17 @@ func TestRegister_UnsupportedOS(t *testing.T) {
 	_ = err
 }
 
-func TestStop_UnsupportedOS(t *testing.T) {
-	_ = Stop()
+func TestStop_WithWorkDir(t *testing.T) {
+	_ = Stop("/fake/workdir")
 }
 
 func TestIsRunning(t *testing.T) {
-	_ = IsRunning()
+	_ = IsRunning("/fake/workdir")
+}
+
+func TestServiceName(t *testing.T) {
+	name := serviceName("/some/path")
+	if name == "" {
+		t.Error("serviceName should return non-empty string")
+	}
 }