+++ internal/service/service_windows.go
@@ -4,7 +4,7 @@ package service
 
 import "fmt"
 
-func registerLaunchd(execPath string) error {
+func registerLaunchd(execPath, workDir string) error {
 	return fmt.Errorf("launchd is not available on Windows")
 }
 
@@ -12,18 +12,30 @@ func registerSystemd(execPath, workDir string) error {
 	return fmt.Errorf("systemd is not available on Windows")
 }
 
-func startLaunchd() error {
+func startLaunchd(workDir string) error {
 	return fmt.Errorf("launchd is not available on Windows")
 }
 
-func startSystemd() error {
+func startSystemd(workDir string) error {
 	return fmt.Errorf("systemd is not available on Windows")
 }
 
-func stopLaunchd() error {
+func stopLaunchd(workDir string) error {
 	return fmt.Errorf("launchd is not available on Windows")
 }
 
-func stopSystemd() error {
+func stopSystemd(workDir string) error {
 	return fmt.Errorf("systemd is not available on Windows")
+}
+
+func isRunningLaunchd(workDir string) bool {
+	return false
+}
+
+func isRunningSystemd(workDir string) bool {
+	return false
+}
+
+func listServices() ([]ServiceInfo, error) {
+	return nil, fmt.Errorf("service listing not supported on Windows")
 }